pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-signal-agent"
version = "0.2.0"
description = "Signal messenger bridge for NanoClaw AI agent runtime"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["nanoclaw", "signal", "agent", "ai", "assistant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Communications :: Chat",
]

dependencies = [
    "claude-code-sdk>=0.1.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "questionary>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "cryptography>=41.0.0",
]

[project.scripts]
nanoclaw = "nanoclaw.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/nanoclaw"]
